Not necessary to use uint64_t for number of blocks owned by GC, size_t suffices.  Micro-optimization.  (r+me)

+++ MMgc/GC.cpp
@@ -402,7 +402,7 @@ namespace MMgc
 	}
 #endif
 
-	uint64_t GCPolicyManager::blocksOwnedByGC() {
+	size_t GCPolicyManager::blocksOwnedByGC() {
 		return blocksOwned;
 	}
